Implement Position entity and integrate with Team creation process

+++ internal/models/vacancy.go
@@ -2,11 +2,11 @@ package models
 
 type Vacancy struct {
 	Role    Role
-	Vacancy int
+	Vacancy int8
 }
 
 // NewVacancy creates a new Vacancy instance
-func NewVacancy(role Role, vacancy int) Vacancy {
+func NewVacancy(role Role, vacancy int8) Vacancy {
 	return Vacancy{
 		Role:    role,
 		Vacancy: vacancy,
